epubview: font size and invert colors buttons
I've added a new menu to control font options. In this menu I've added the
font size buttons and another button to invert colors.

The invert colors assume that the background is light and the text is dark
and when it's inverted sets the background to black and the text to white.

The menu is ugly now, if we can create some buttons or widgets in this
menu, it will look a lot better.